// src/types/expense-income.ts
// Generated by https://quicktype.io

export type ExpenseIncomeEntity = {
  id: number;
  paymentType: PaymentType;
  status: PaymentStatus;
  issueDate: String;
  dueDate: string;
  categoryId: number;
  categoryDescription: string;
  amount: number;
  history: string;
  bankAccountId?: number;
  bankAccountDescription?: string;
  paymentDate?: String;
  paidAmount?: number;
  overdue: boolean;
  feesAndInterest?: number;
  dueDay: number;
};

// Generated by https://quicktype.io

export type ExpenseIncomeSummary = {
  pendingAmount: number;
  totalAmount: number;
  list: ExpenseIncomeEntity[];
}

export enum PaymentType {
  Expense = "EXPENSE",
  Income = "INCOME",
}

export enum PaymentStatus {
  Paid = "PAID",
  Pending = "PENDING",
}